+++ workspace_declaration_test.go
@@ -1,14 +1,8 @@
 package langd
 
 import (
-	"bytes"
-	"fmt"
 	"go/token"
-	"os"
 	"testing"
-
-	"github.com/object88/langd/log"
-	"golang.org/x/tools/go/buildutil"
 )
 
 func Test_Workspace_Declaration_Package_Const(t *testing.T) {
@@ -38,7 +32,7 @@ func Test_Workspace_Declaration_Package_Const(t *testing.T) {
 		Line:     3,
 		Column:   3,
 	}
-	test(t, w, usagePosition, declPosition)
+	testDeclaration(t, w, usagePosition, declPosition)
 }
 
 func Test_Workspace_Declaration_Package_Func(t *testing.T) {
@@ -66,7 +60,7 @@ func Test_Workspace_Declaration_Package_Func(t *testing.T) {
 		Line:     2,
 		Column:   7,
 	}
-	test(t, w, usagePosition, declPosition)
+	testDeclaration(t, w, usagePosition, declPosition)
 }
 
 func Test_Workspace_Declaration_Package_Var(t *testing.T) {
@@ -94,7 +88,7 @@ func Test_Workspace_Declaration_Package_Var(t *testing.T) {
 		Line:     2,
 		Column:   6,
 	}
-	test(t, w, usagePosition, declPosition)
+	testDeclaration(t, w, usagePosition, declPosition)
 }
 
 func Test_Workspace_Declaration_Package_Var_CrossFiles(t *testing.T) {
@@ -125,7 +119,7 @@ func Test_Workspace_Declaration_Package_Var_CrossFiles(t *testing.T) {
 		Line:     2,
 		Column:   6,
 	}
-	test(t, w, usagePosition, declPosition)
+	testDeclaration(t, w, usagePosition, declPosition)
 }
 
 func Test_Workspace_Declaration_Package_Var_Shadowed(t *testing.T) {
@@ -154,78 +148,5 @@ func Test_Workspace_Declaration_Package_Var_Shadowed(t *testing.T) {
 		Line:     5,
 		Column:   10,
 	}
-	test(t, w, usagePosition, declPosition)
-}
-
-func workspaceSetup(t *testing.T, startingPath string, packages map[string]map[string]string, expectFailure bool) *Workspace {
-	fc := buildutil.FakeContext(packages)
-	loader := NewLoader(func(l *Loader) {
-		l.context = fc
-	})
-	w := CreateWorkspace(loader, log.CreateLog(os.Stdout))
-	w.log.SetLevel(log.Verbose)
-
-	done := loader.Start()
-	loader.LoadDirectory(startingPath)
-	<-done
-
-	if expectFailure {
-		errCount := 0
-		w.Loader.Errors(func(file string, errs []FileError) {
-			errCount += len(errs)
-		})
-		if errCount == 0 {
-			t.Fatal("Expected errors, but got none\n")
-		}
-	} else {
-		errCount := 0
-		var buf bytes.Buffer
-		w.Loader.Errors(func(file string, errs []FileError) {
-			buf.WriteString(fmt.Sprintf("Loading error in %s:\n", file))
-			for k, err := range errs {
-				buf.WriteString(fmt.Sprintf("\t%02d: %s:%d %s\n", k, err.Filename, err.Line, err.Message))
-			}
-			errCount += len(errs)
-		})
-
-		if errCount != 0 {
-			buf.WriteString(fmt.Sprintf("Total: %d errors\n", errCount))
-			t.Fatal(buf.String())
-		}
-	}
-
-	return w
-}
-
-// func test(t *testing.T, w *Workspace, declOffset, usageOffset int) {
-func test(t *testing.T, w *Workspace, usagePosition, expectedDeclPosition *token.Position) {
-	declPosition, err := w.LocateDeclaration(usagePosition)
-	if err != nil {
-		t.Fatal(err.Error())
-	}
-
-	comparePosition(t, declPosition, expectedDeclPosition)
-}
-
-func comparePosition(t *testing.T, actual, expected *token.Position) {
-	if actual == nil {
-		t.Fatalf("actual is nil")
-	}
-
-	if !actual.IsValid() {
-		t.Fatalf("Returned position '%s' is not valid", actual.String())
-	}
-
-	if actual.Filename != expected.Filename {
-		t.Fatalf("Incorrect filename: got '%s', expected '%s'", actual.Filename, expected.Filename)
-	}
-
-	if actual.Line != expected.Line {
-		t.Fatalf("Incorrect line: got %d, expected %d", actual.Line, expected.Line)
-	}
-
-	if actual.Column != expected.Column {
-		t.Fatalf("Incorrect column: got %d, expected %d", actual.Column, expected.Column)
-	}
-
+	testDeclaration(t, w, usagePosition, declPosition)
 }
