+++ workspace_imports_test.go
@@ -1,13 +1,5 @@
 package langd
 
-import (
-	"os"
-	"testing"
-
-	"github.com/object88/langd/log"
-	"golang.org/x/tools/go/buildutil"
-)
-
 const identImportsTestProgram1 = `package foo
 
 import (
@@ -57,42 +49,3 @@ func CountCall(source string) {
 // 		t.Fatalf("Returned nil from LocateReferences")
 // 	}
 // }
-
-func setup2(t *testing.T) *Workspace {
-	packages := map[string]map[string]string{
-		"foo": map[string]string{
-			"foo.go": identImportsTestProgram1,
-		},
-		"bar": map[string]string{
-			"bar.go": identImportsTestProgram2,
-		},
-	}
-
-	fc := buildutil.FakeContext(packages)
-	loader := NewLoader(func(l *Loader) {
-		l.context = fc
-	})
-	w := CreateWorkspace(loader, log.CreateLog(os.Stdout))
-	w.log.SetLevel(log.Verbose)
-
-	done := loader.Start()
-	loader.LoadDirectory("/go/src/foo")
-	<-done
-
-	errCount := 0
-	w.Loader.Errors(func(file string, errs []FileError) {
-		if errCount == 0 {
-			t.Errorf("Loading error in %s:\n", file)
-		}
-		for k, err := range errs {
-			t.Errorf("\t%d: %s\n", k, err.Message)
-		}
-		errCount++
-	})
-
-	if errCount != 0 {
-		t.Fatalf("Found %d errors", errCount)
-	}
-
-	return w
-}
